server/requirements.txt
alembic==1.13.2
amqp==5.2.0
aniso8601==9.0.1
attrs==23.2.0
Authlib==1.3.1
Babel==2.15.0
bcrypt==4.2.0
billiard==4.2.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.40.0
colorama==0.4.6
cryptography==43.0.0
Flask==3.0.3
Flask-Admin==1.6.1
flask-babel==4.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-DebugToolbar==0.15.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
Pillow==9.5.0
prompt_toolkit==3.0.47
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.19.1
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
WTForms==3.1.2
